Move RatingPainter to it's own file
